pyproject.toml
[tool.poetry]
name = "qlbbl-bot"
version = "0.1.0"
description = ""
authors = ["Qlbbl1111 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
Flask = "^2.0.2"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
